Correct BaseId as MappedSuperclass
