+++ elide-integration-tests/src/test/java/example/Post.java
@@ -12,12 +12,10 @@ import com.yahoo.elide.annotation.ReadPermission;
 import com.yahoo.elide.annotation.UpdatePermission;
 import com.yahoo.elide.security.Role;
 
-import javax.persistence.Entity;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
 import java.util.HashSet;
 import java.util.Set;
+
+import javax.persistence.Entity;
 
 /**
  * Post test bean.
@@ -32,7 +30,6 @@ import java.util.Set;
 public class Post extends BaseId {
     private String title;
     private int created;
-    private long id;
     private final Set<Parent> spouses = new HashSet<>();
 
     @ReadPermission(all = { Role.NONE.class }) public transient boolean init = false;
@@ -40,17 +37,6 @@ public class Post extends BaseId {
     public void doInit() {
         init = true;
     }
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public long getId() {
-        return id;
-    }
-
-    public void setId(long id) {
-        this.id = id;
-    }
-
 
     public String getTitle() {
         return this.title;
